projects/drone-vision/lambda/requirements.txt
# projects/drone-vision/lambda/requirements.txt
# --------------------------------------------
#  Lightweight CPU-only runtime for ONNX   (1.16.3 is last wheel with
#  manylinux2014 tags that work out-of-the-box on Lambda Python 3.11)
onnxruntime==1.16.3
numpy==1.26.4
pillow==11.3.0

# Vision utilities (headless OpenCV keeps size down)
opencv-python-headless>=4.10.0.82

# Heat-map extras
scipy>=1.11                # Gaussian blur / KDE
redis>=5.0                 # optional multi-frame grid store
boto3>=1.34                # S3 client for GeoJSON uploads
matplotlib>=3.8            # extra colour-maps (viridis, plasma, …)
